+++ lib/puppet/functions/extended_stdlib/drop.rb
@@ -3,7 +3,7 @@
 # Returns an Array containing all but the first num element of the array, where num is a non-negative Integer
 Puppet::Functions.create_function(:'extended_stdlib::drop') do
   # @param the_array The array to remove the first num elements from.
-  # @param num The number of first elements to remove from the array.
+  # @param num The number of first elements to remove from the array. This must be a non-negative Integer.
   # @return [Array] Returns the array with the first num elements removed.
   # @example Remove the first 0, 1, or 2 elements from an array, and return that array.
   #   drop([0, 1, 2, 3, 4, 5], 0) => [0, 1, 2, 3, 4, 5]