+++ lib/puppet/functions/extended_stdlib/start_with.rb
@@ -21,8 +21,7 @@ Puppet::Functions.create_function(:'extended_stdlib::start_with') do
   def start_with(a_string, prefixes)
     # any?(Regexp) coincidentally also only ruby >= 2.5
     if prefixes.any? { |prefix| prefix.instance_of?(Regexp) } && Gem::Version.new(RUBY_VERSION) < Gem::Version.new('2.5')
-      Puppet.warning('Regular expression prefixes are not supported in Ruby prior to version 2.5.')
-      false
+      raise Puppet::Error, 'Regular expression prefixes are not supported in Ruby prior to version 2.5.'
     else
       a_string.start_with?(*prefixes)
     end